ManualTradeManager: Initial commit
This is an idea I'm working on to manage a way of trading that has worked well for me recently, entering small with stops outside market structure and scaling up if it goes in the right direction.

So far it only makes the stop break even (and a little bit plus so there won't be tiny losses on the account messing up the stats if it gets hit).

Once and if this works reliably there could be a point to trying to automate the entry decisions as well to get a fully automated strategy.